Attempt to use the garden-cli library
I need a better library than getopt to parse options, so that it doesn't depend
on the process being run but so I can set it arbitrarily; then I will be able to
test the code execution paths on the program all return the right things.

This library is not useful because the Cli::writeUsage method is protected, but
it won't allow me to require ONE OF file or write_table. I can't use the library
without massively breaking encapsulation.